Make more improvements to the database/tables

+++ sql/make-film-incrementor.sql
@@ -1,6 +1,6 @@
 -- Add the new column
 ALTER TABLE `films`
-ADD COLUMN `incrementor` INT(12) UNSIGNED NOT NULL DEFAULT '0' AFTER `edit`;
+ADD COLUMN `incrementor` SMALLINT(5) UNSIGNED NOT NULL DEFAULT '0' AFTER `edit`;
 
 -- Add an index to improve lookup
 ALTER TABLE `films`
